+++ brick_breaker/src/vertex_buffer.h
@@ -1,9 +1,11 @@
 #ifndef VERTEX_BUFFER_H
 #define VERTEX_BUFFER_H
 
+#include <cstdint>
+
 class VertexBuffer {
 private:
-	unsigned int obj_id;
+	uint32_t obj_id;
 
 public:
 	VertexBuffer();
@@ -12,7 +14,7 @@ public:
 	void bind() const;
 	void unbind() const;
 
-	void buffer_data(const void* data, size_t size, unsigned int usage) const;
+	void buffer_data(const void* data, uint32_t size, uint32_t usage) const;
 };
 
 #endif // !VERTEX_BUFFER_H
